hw2 finished

+++ hw2/hw2/PercolationStats.java
@@ -4,13 +4,13 @@ import edu.princeton.cs.algs4.StdRandom;
 import edu.princeton.cs.algs4.StdStats;
 
 public class PercolationStats{
-    private int[] arrOfThreshold;
+    private double[] arrOfThreshold;
 
     public PercolationStats(int N, int T, PercolationFactory pf){
     if(N<=0||T<=0){
         throw new java.lang.IllegalArgumentException();
     }
-    arrOfThreshold=new int[T];
+    arrOfThreshold=new double[T];
     int randomRow;
     int randomCol;
     for (int i=0;i<T;i+=1){
@@ -20,7 +20,7 @@ public class PercolationStats{
             randomCol=StdRandom.uniform(N);
             test.open(randomRow,randomCol);
         }
-        arrOfThreshold[i]=test.numberOfOpenSites()/(N*N);
+        arrOfThreshold[i]=(double)test.numberOfOpenSites()/(N*N);
     }
     }   // perform T independent experiments on an N-by-N grid
     public double mean(){
